updates according to the pull request
1-rename item file to homepage
2-rename actorsCard to ActorList
3-create a component for movieList mapping
