+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.16.5",
     "@mui/material": "^5.16.0",
     "@popperjs/core": "^2.11.8",
     "axios": "^1.7.2",
